+++ src/Context/SchoolProvider.js
@@ -1,24 +1,44 @@
 import React, { createContext, useState, useEffect, useRef, useContext, useCallback } from 'react';
 import RestService from "../Services/RestService"
+import { useNavigationContext } from './NavigationProvider';
 
 export const SchoolContext = createContext();
 
 export const useSchoolContext = () => useContext(SchoolContext);
 
-export const SchoolProvider = ({ children, value }) => {
+const emptyDocument = {
+    budget: 0,
+    cashAdvance: 0
+}
+
+// Initialize current date to get current month and year
+const currentDate = new Date();
+const currentMonth = currentDate.toLocaleString('default', { month: 'long' }); // Get full month name
+const currentYear = currentDate.getFullYear().toString(); // Get full year as string
+
+const months = [
+    'January', 'February', 'March', 'April', 'May', 'June',
+    'July', 'August', 'September', 'October', 'November', 'December'
+];
+
+const years = [
+    '2021', '2022', '2023', '2024'
+];
+
+export const SchoolProvider = ({ children }) => {
     // Set initial state for month and year using current date
-    const {
-        currentMonth, currentYear,
-        currentSchool, setCurrentSchool,
-        currentDocument, setCurrentDocument,
-        month, setMonth,
-        year, setYear,
-        months, years,
-        isAdding, setIsAdding,
-        addOneRow, setAddOneRow,
-        reload, setReload,
-        fetchDocumentData, val
-    } = value;
+    const { currentSchool } = useNavigationContext();
+
+    // Set initial state for month and year using current date
+    const [month, setMonth] = useState(currentMonth);
+    const [year, setYear] = useState(currentYear);
+
+    const [isAdding, setIsAdding] = useState(false);
+    const [addOneRow, setAddOneRow] = useState(false);
+
+    const [reload, setReload] = useState(false);
+
+    const [currentDocument, setCurrentDocument] = useState(null);
     const [lr, setLr] = useState([]);
 
     const monthIndex = months.indexOf(currentMonth);
@@ -26,6 +46,26 @@ export const SchoolProvider = ({ children, value }) => {
 
     const prevMonthRef = useRef(monthIndex === 0 ? 11 : monthIndex);
     const prevYearRef = useRef(monthIndex === 0 ? (yearIndex === 0 ? years.length - 1 : yearIndex - 1) : yearIndex);
+
+    const fetchDocumentData = useCallback(async () => {
+        try {
+            if (currentSchool) {
+                const getDocument = await RestService.getDocumentBySchoolIdYearMonth(
+                    currentSchool?.id,
+                    year,
+                    month
+                );
+
+                if (getDocument) {
+                    setCurrentDocument(getDocument);
+                } else {
+                    setCurrentDocument(emptyDocument);
+                }
+            }
+        } catch (error) {
+            console.error('Error fetching document:', error);
+        }
+    }, [currentSchool, setCurrentDocument, year, month]);
 
     const fetchLrByDocumentId = async (id) => {
         try {
@@ -78,18 +118,19 @@ export const SchoolProvider = ({ children, value }) => {
 
     useEffect(() => {
         console.log("SchoolProvider useEffect: update lr");
-        if (val === 0) { // if LR tab is selected
-            updateLr();
-        }
+        //if (val === 0) { // if LR tab is selected
+        //updateLr();
+        //}
+        fetchDocumentData();
 
-    }, [month, year, currentDocument, val, updateLr]); // Run effect only on mount and unmount*/
+    }, [month, year, currentSchool, fetchDocumentData]); // Run effect only on mount and unmount*/
 
     return (
         <SchoolContext.Provider value={{
             prevMonthRef, prevYearRef, month, setMonth, year, setYear, months, years,
             lr, setLr, fetchLrByDocumentId, setCurrentDocument, currentDocument,
             displayFields, isAdding, setIsAdding, addOneRow, setAddOneRow, updateLr, fetchDocumentData,
-            currentSchool, setCurrentSchool, reload, setReload
+            currentSchool, reload, setReload
         }}>
             {children}
         </SchoolContext.Provider>
